// src/Cumplimiento/interfaces/getCumplimiento.interfaces.ts
export interface DiaCal {
    id_header:      number;
    dia:            Date | string;
    servicio:       string;
    eco:            number;
    op_cred:        number;
    op_tipo:        string;
    op_estado:      string;
    sistema:        string;
    id_jornada?:    number;
    turno:          1|2|3;
    hr_ini_t:       string;
    lug_ini_cc:     string;
    hr_ini_cc?:     string;
    hr_ter_cc?:     string;
    lug_ter_cc:    string;
    hr_ter_mod?:    string;
    hr_ter_t:       string;
}